Move ColumnHeaderEditor
